+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ### vNEXT
 
+- **core**: Fix "Cannot convert object to primitive value" error for ChangeType.InputFieldDefaultValueChanged
+
 ### v3.2.0
 
 - **action**: Use GitHub REST API to retrieve the full blob content if schema is truncated [#2131](https://github.com/kamilkisiela/graphql-inspector/pull/2131)
 - **action**: Diffing a graphql endpoint with a local .json schema file results in an error, fixed in [#2090](https://github.com/kamilkisiela/graphql-inspector/pull/2090)
 - **core**: Fix comparison of non-nullable fields [#2122](https://github.com/kamilkisiela/graphql-inspector/pull/2122)
 - **core**: Fix removal of a deprecated input field [#2122](https://github.com/kamilkisiela/graphql-inspector/pull/2122)
-
 
 ### v3.1.4